requirements.txt
altair==4.2.0
numpy
pandas
plotly-express
streamlit==1.8.1
